net/freeradius3: Remove RUBY
As per comments in the source this module still is unstable and
(other than majority of experimental) conetinas a warning about
leaking memory and discourages maintainers to make use of it.

Keeping in mind that the module is present since quite a couupl of
years and contains this warning since the start of the v3.0 releases.

Thus I'm inclinded to rip this option out of the port if even upstream
warns about its usage.