net/freeradius3: Remove extra handling for amd64
This is likely a historical relic taken from net/freeradius2 which
was the starting point for this port.

It was introduced in 2004 with:

    Fix package build on AMD64

    PR:         72335
    Submitted by:       Andrew Thompson <andy@fud.org.nz>
    Approved by:        portmgr

And last removed when 2.2.5 was the "version du jour" july 2014 with:

    - Convert to USES=libtool
    - Add USE_AUTOTOOLS=libtoolize (and aclocal) because the bundled libtool
      is too old to work with devel/libltdl
    - Move several variable definitions before inclusion of bsd.port.options.mk
    - Remove use of UNIQUENAME

net/freeradius3 started as a copy of net/freeradius2 which was made in 2013 when
this was still present there.

Unfortunately no release news on FreeRADIUS.org indicates around that timeframe
something FreeBSD related on that topic.

Current opinion: The technical need as workaround for FreeBSD amd64 isn't
there on any currently supported version of FreeBSD with this version of
FreeRADIUS 3.

Though: If we should enable it globally is another question to be thought of.