net/freeradius3: Simplify LDAP, drops EDIR
Actually EDIR was faulty (as I did) and wouldn't have worked it previously.

eDirectory Support doesn't induce another dependency and is enabled
with rlm_ldap. If we enable it with LDAP support it simplifies the port
and the people don't have to rebuild for extra support if they eventually
have this type of LDAP directory service.

I don't have a way to check whether eDirectory support really works.
It's part of rlm_ldap which itself loads but I can't connect to an actual
server here.